requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.4
django-dotenv==1.4.2
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
kombu==5.3.7
packaging==24.0
prompt-toolkit==3.0.43
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
